requirements.txt
groq==0.9.0
langchain==0.2.16
langchain-community==0.2.16
langchain-huggingface==0.0.3
faiss-cpu==1.8.0
sentence-transformers==3.0.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
python-dotenv==1.0.1
rouge==1.0.1
tiktoken==0.7.0
tqdm==4.66.4
colorama==0.4.6
tabulate==0.9.0
httpx==0.27.0